+++ R/semtree.constraints.R
@@ -11,8 +11,6 @@
 #' @param local.invariance Vector of parameter names that are locally equal,
 #' that is, they are assumed to be equal when assessing a local split but
 #' allowed to differ subsequently.
-#' @param global.invariance Vector of parameter names that are globally equal,
-#' that is, estimated only once and then fixed in the tree.
 #' @param focus.parameters Vector of parameter names that exclusively are
 #' evaluated for between-group differences when assessing split candidates. If
 #' NULL all parameters add to the difference.
@@ -22,11 +20,10 @@
 #' U. (2013). Structural equation model trees. \emph{Psychological Methods},
 #' 18(1), 71-86.
 #' @export
-semtree.constraints <- function(local.invariance=NULL, global.invariance=NULL, focus.parameters=NULL)
+semtree.constraints <- function(local.invariance=NULL, focus.parameters=NULL)
 {
   ret <- list()
   ret$local.invariance <- local.invariance
-  ret$global.invariance <- global.invariance
   ret$focus.parameters <- focus.parameters
   class(ret) <- "semtree.constraints"
   
@@ -37,6 +34,5 @@ semtree.constraints <- function(local.invariance=NULL, global.invariance=NULL, f
 print.semtree.constraints <- function(x, ...) {
   cat("semtree constraints\n")
   cat(" Local invariance ",paste0(x$local.invariance),"\n")
-  cat(" Global invariance ",paste0(x$global.invariance),"\n")
   cat(" Focus parameters ",paste0(x$focus.parameter),"\n")
 }